chore: port login and register to features
