--- apps/admin/src/components/login/views/style.module.css
@@ -1,61 +1,0 @@
-.wrapper {
-  display: flex;
-  justify-content: space-around;
-  /* box-shadow: 0 0 10px 3px #eee; */
-  margin: 0 20vw;
-  padding: 0 0px;
-}
-.leftBox {
-  width: 40%;
-  align-items: center;
-  display: flex;
-  justify-content: center;
-  font-size: 16px;
-  opacity: 0.8;
-  color: #686767;
-  font-weight: 400;
-  font-family: system-ui;
-  text-transform: uppercase;
-  letter-spacing: 8px;
-  gap: 16px;
-  flex-direction: column;
-  border-right: 1px solid rgb(var(--color-border));
-}
-.rightBox {
-  flex: 1;
-}
-.forms {
-  padding: 20px;
-  padding: 30px;
-  background: rgb(var(--section-bg));
-  border: 1px solid rgb(var(--color-border));
-  border-radius: 2px;
-  overflow: hidden;
-  margin: auto;
-  flex: 1;
-  text-align: center;
-}
-
-@media (max-width: 950px) {
-  .forms {
-    padding: 16px;
-    border: none;
-  }
-  .leftBox {
-    display: none;
-  }
-  .wrapper {
-    width: auto;
-    flex-direction: column;
-  }
-}
-
-@media (max-width: 700px) {
-  .wrapper {
-    margin: 0 0vw;
-    margin: 20px;
-  }
-  .leftBox {
-    width: auto;
-  }
-}
