+++ apps/admin/src/features/login/index.ts
@@ -1,0 +1,3 @@
+export { Feature } from "./feature";
+export { Logo } from "./logo";
+export { SocialLogin } from "./views/social-login";
